requirements.txt
absl-py==1.2.0
astor==0.8.1
cached-property==1.5.2
certifi==2021.5.30
charset-normalizer==2.0.12
colorama==0.4.5
cycler==0.11.0
dataclasses==0.8
gast==0.2.2
google-pasta==0.2.0
grpcio==1.47.0
h5py==2.10.0
idna==3.3
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.2.2
mock==4.0.3
numpy==1.19.5
opencv-python-headless==4.1.1.26
opt-einsum==3.3.0
Pillow==8.4.0
protobuf==3.19.4
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
scipy==1.5.4
six==1.16.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.2
termcolor==1.1.0
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.11
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.14.1
zipp==3.6.0
